requirements.txt
aioredis==2.0.1
alembic==1.8.1
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
bcrypt==3.2.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cleo==1.0.0a4
click==8.1.3
crashtest==0.3.1
cryptography==37.0.2
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fastapi==0.78.0
fastapi-redis-cache==0.2.5
greenlet==1.1.2
h11==0.13.0
hiredis==2.0.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
inflect==5.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
orjson==3.7.2
packaging==21.3
passlib==1.7.4
pptree==3.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
pylev==1.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
python-ulid==1.1.0
PyYAML==6.0
redis==4.3.3
requests==2.28.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
sqlacodegen==2.3.0
SQLAlchemy==1.4.37
starlette==0.19.1
types-redis==4.2.6
types-six==1.16.15
typing-extensions==4.2.0
ujson==5.3.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
wrapt==1.14.1
zipp==3.8.1
